backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.32.0,<0.33.0
motor>=3.3.0,<4.0.0
pydantic>=2.5.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
httpx>=0.27.1,<0.28.0
aiohttp>=3.8.0,<4.0.0
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-openai>=0.2.0,<0.3.0
langchain-anthropic>=0.2.0,<0.3.0
websockets>=12.0,<13.0
langextract>=0.1.0
google-generativeai>=0.3.0

# MCP (Model Context Protocol) SDK - compatible with anyio>=4.5
mcp>=1.24.0,<2.0.0

# FastMCP for simplified MCP server creation
fastmcp>=0.1.0

# Google API libraries for Gmail MCP server
google-auth>=2.0.0,<3.0.0
google-auth-oauthlib>=1.0.0,<2.0.0
google-auth-httplib2>=0.2.0,<1.0.0
google-api-python-client>=2.0.0,<3.0.0
googleapis-common-protos>=1.60.0,<2.0.0

# File upload and parsing
python-docx>=1.1.0,<2.0.0
python-magic>=0.4.27,<0.5.0

# Configuration management
PyYAML>=6.0,<7.0
